src/chart-builder/pyproject.toml
[project]
name = "chart-builder"
version = "1.0.0"
description = "Python packgage that installs helm charts to Spreetail's AKS Clusters."
authors = [
    {name = "Mason Huemmer", email = "mason.huemmer@spreetail.com"},
]
dependencies = [
    "pyyaml>=5.3.1",
    "azure-identity>=1.10.0",
    "azure-mgmt-containerservice>=20.3.0",
    "kubernetes>=24.2.0",
    "azure-mgmt-resource>=21.1.0",
    "azure-mgmt-subscription>=3.1.1",
    "datadog-api-client>=2.3.0",
    "requests>=2.28.1",
    "rich>=12.6.0",
]
requires-python = ">=3.7"
license = {text = "MIT"}
[project.optional-dependencies]

[tool]
[tool.pdm]
[tool.pdm.scripts]
_.env_file = ".env.local"
app = "python chart-builder"
lint = "pylint chart --reports y --recursive y --exit-zero"
test = "coverage run -m pytest"
[tool.pdm.overrides]
azure-identity = "1.10.0"
azure-mgmt-containerservice = "20.3.0"
kubernetes = "24.2.0"

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.1.3",
    "pytest-cov>=4.0.0",
    "pylint>=2.13.9",
]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
